requirements.txt
livekit-agents>=0.12.1
livekit-plugins-openai>=0.10.9
livekit-plugins-deepgram>=0.6.13
python-dotenv~=1.0
livekit-plugins-turn-detector==0.3.4
livekit-plugins-elevenlabs==0.7.9
livekit-plugins-silero==0.7.4
